requirements.txt
fastapi
uvicorn
weasyprint
jinja2
python-dotenv
pydantic
email-validator
pymongo